requirements.txt
appdirs==1.4.4
asttokens==2.4.1
attrs==21.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
Brotli==1.0.9
chardet==4.0.0
checkpoint_schedules==1.0.4
coverage==7.9.1
cycler==0.11.0
Cython==0.29.32
decorator==4.4.2
-e git+https://github.com/dolfin-adjoint/dolfin-adjoint@388f94be89a8cd43a40dc465da2e763f8413be01#egg=dolfin_adjoint
exceptiongroup==1.2.1
executing==2.0.1
fenics-dijitso @ file:///src/dijitso
fenics-dolfin @ file:///src/dolfin/python
fenics-ffc @ file:///src/ffc
fenics-fiat @ file:///src/fiat
fenics-ufl-legacy @ file:///src/ufl-legacy
flake8==7.3.0
fonttools==4.29.1
fs==2.4.12
future==1.0.0
gast==0.5.2
gmsh==4.12.2.dev1
h5py==3.11.0
html5lib==1.1
iniconfig==2.0.0
ipopt==1.0.3
ipython==8.24.0
jedi==0.19.1
kiwisolver==1.3.2
lxml==4.8.0
lz4==3.1.3+dfsg
markdown-it-py==3.0.0
matplotlib==3.5.1
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
meshio==5.3.5
moola @ git+https://github.com/funsim/moola@79ffc8d86dff383762063970358b8aebfe373896
mpi4py==3.1.6
mpmath==0.0.0
numpy==1.24.4
olefile==0.46
packaging==21.3
pandas==2.3.1
parso==0.8.4
petsc4py @ file:///usr/local/petsc/src/binding/petsc4py
pexpect==4.9.0
Pillow==9.0.1
pkgconfig==1.5.5
pluggy==1.5.0
ply==3.11
prompt_toolkit==3.0.44
ptyprocess==0.7.0
pure-eval==0.2.2
pyadjoint-ad==2025.4.0
pybind11==2.10.1
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.18.0
pyparsing==2.4.7
pytest==8.2.1
python-dateutil==2.9.0.post0
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
rich==13.7.1
ROL==0.0.16
roltrilinos==0.0.9
scipy==1.8.0
six==1.16.0
slepc4py @ file:///usr/local/slepc/src/binding/slepc4py
soupsieve==2.3.1
stack-data==0.6.3
sympy==1.9
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2025.2
ufoLib2==0.13.1
unicodedata2==14.0.0
wcwidth==0.2.13
webencodings==0.5.1
